refactor: 패키지 이동 및 Spring EL 변환 로직 추가
