+++ cinema-adapter/src/main/java/com/cinema/adapter/in/web/dto/request/NowPlayingMovieRequest.java
@@ -1,6 +1,6 @@
 package com.cinema.adapter.in.web.dto.request;
 
-import com.cinema.adapter.in.web.validation.EnumValue;
+import com.cinema.adapter.in.web.annotation.EnumValue;
 import com.cinema.domain.enums.MovieGenre;
 import jakarta.validation.constraints.Size;
