+++ cinema-adapter/src/main/java/com/cinema/adapter/in/web/validation/EnumValidator.java
@@ -1,5 +1,6 @@
 package com.cinema.adapter.in.web.validation;
 
+import com.cinema.adapter.in.web.annotation.EnumValue;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
