+++ cinema-application/src/main/java/com/cinema/application/service/CreateReservationService.java
@@ -26,7 +26,6 @@ public class CreateReservationService implements CreateReservationUseCase {
     private final ApplicationEventPublisher eventPublisher;
 
     @Override
-    @Transactional
     public List<CreateReservationResult> createReservation(CreateReservationCommand command) {
         TicketReservation reservation = ReservationMapper.toDomain(command);
